Build the mathlib of 3rd/mathlib, with the fixes prepare.cmake applied, in the three builds rather than downloading it GAOL can be brought into a project by FetchContent (tests/fetch_content); an installed mathlib is used only when asked for

+++ 3rd/mathlib/src/mpsqrt.c
@@ -82,9 +82,9 @@ void mpsqrt(mp_no *x, mp_no *y, int p) {
 /* with the relative error bounded by 2**-51.              */
 /***********************************************************/
 double fastiroot(double x) {
-  union {long i[2]; double d;} p,q;
+  union {int i[2]; double d;} p,q;
   double y,z, t;
-  long n;
+  int n;
   static const double c0 = 0.99674, c1 = -0.53380, c2 = 0.45472, c3 = -0.21553;
   
   p.d = x;
